fix: module imports refactor
