+++ apps/enrolla-server/src/app.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AuthzModule } from './authz/authz.module';
-import { PrismaService } from './prisma.service';
 import { FeaturesModule } from './features/features.module';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { PackagesModule } from './packages/packages.module';
@@ -15,6 +14,7 @@ import { CustomersModule } from './customers/customers.module';
 import { IntegrationsModule } from './integrations/integrations.module';
 import { OrganizationsModule } from './organizations/organizations.module';
 import { ConfigurationsModule } from './configurations/configurations.module';
+import { PrismaModule } from './prisma/prisma.module';
 
 @Module({
   imports: [
@@ -28,6 +28,7 @@ import { ConfigurationsModule } from './configurations/configurations.module';
       },
     }),
     AuthzModule,
+    PrismaModule,
     FeaturesModule,
     PackagesModule,
     FeatureInstancesModule,
@@ -36,7 +37,6 @@ import { ConfigurationsModule } from './configurations/configurations.module';
     OrganizationsModule,
     ConfigurationsModule,
   ],
-  providers: [PrismaService],
   controllers: [AppController],
 })
 export class AppModule {}
