+++ apps/enrolla-server/src/configurations/configurations.module.ts
@@ -3,5 +3,6 @@ import { ConfigurationsService } from './configurations.service';
 
 @Module({
   providers: [ConfigurationsService],
+  exports: [ConfigurationsService],
 })
 export class ConfigurationsModule {}
