+++ apps/enrolla-server/src/customers/customers.module.ts
@@ -1,21 +1,13 @@
 import { Module } from '@nestjs/common';
 import { CustomersService } from './customers.service';
 import { CustomersResolver } from './customers.resolver';
-import { OrganizationsService } from '../organizations/organizations.service';
-import { PrismaService } from '../prisma.service';
-import { ConfigurationsService } from '../configurations/configurations.service';
-import { FeatureInstancesService } from '../feature-instances/feature-instances.service';
 import { PackagesModule } from '../packages/packages.module';
+import { OrganizationsModule } from '../organizations/organizations.module';
+import { FeatureInstancesModule } from '../feature-instances/feature-instances.module';
 
 @Module({
-  providers: [
-    CustomersResolver,
-    CustomersService,
-    OrganizationsService,
-    ConfigurationsService,
-    PrismaService,
-    FeatureInstancesService,
-  ],
-  imports: [PackagesModule],
+  providers: [CustomersResolver, CustomersService],
+  imports: [PackagesModule, OrganizationsModule, FeatureInstancesModule],
+  exports: [CustomersService],
 })
 export class CustomersModule {}
