+++ apps/enrolla-server/src/packages/packages.module.ts
@@ -1,20 +1,14 @@
 import { Module } from '@nestjs/common';
 import { PackagesService } from './packages.service';
 import { PackagesController } from './packages.controller';
-import { PrismaService } from '../prisma.service';
 import { PackagesResolver } from './packages.resolver';
-import { FeatureInstancesService } from '../feature-instances/feature-instances.service';
 import { FeaturesModule } from '../features/features.module';
+import { FeatureInstancesModule } from '../feature-instances/feature-instances.module';
 
 @Module({
   controllers: [PackagesController],
-  providers: [
-    PackagesResolver,
-    PackagesService,
-    PrismaService,
-    FeatureInstancesService,
-  ],
-  imports: [FeaturesModule],
+  providers: [PackagesResolver, PackagesService],
+  imports: [FeaturesModule, FeatureInstancesModule],
   exports: [PackagesService],
 })
 export class PackagesModule {}
