+++ apps/enrolla-server/src/tenants/tenants.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
-import { PrismaService } from '../prisma.service';
 import { TenantsController } from './tenants.controller';
 import { TenantsService } from './tenants.service';
 
 @Module({
   controllers: [TenantsController],
-  providers: [TenantsService, PrismaService],
+  providers: [TenantsService],
 })
 export class TenantsModule {}
